add SurfaceMesh to opengl_surface.py
