+++ manim/mobject/types/opengl_surface.py
@@ -2,7 +2,8 @@ import moderngl
 import numpy as np
 
 from ...constants import *
-from ...mobject.opengl_mobject import OpenGLMobject
+from ...mobject.mobject import Mobject
+from ...mobject.types.vectorized_mobject import VMobject, VGroup
 from ...utils.bezier import integer_interpolate, interpolate
 from ...utils.color import *
 from ...utils.config_ops import _Data, _Uniforms
@@ -11,7 +12,7 @@ from ...utils.iterables import listify
 from ...utils.space_ops import normalize_along_axis
 
 
-class OpenGLSurface(OpenGLMobject):
+class Surface(Mobject):
     shader_dtype = [
         ("point", np.float32, (3,)),
         ("du_point", np.float32, (3,)),
@@ -128,7 +129,7 @@ class OpenGLSurface(OpenGLMobject):
         return normalize_along_axis(normals, 1)
 
     def pointwise_become_partial(self, smobject, a, b, axis=None):
-        assert isinstance(smobject, OpenGLSurface)
+        assert isinstance(smobject, Surface)
         if axis is None:
             axis = self.prefered_creation_axis
         if a <= 0 and b >= 1:
@@ -298,7 +299,7 @@ class OpenGLSurface(OpenGLMobject):
         return self
 
 
-class OpenGLSurfaceGroup(OpenGLSurface):
+class SurfaceGroup(Surface):
     def __init__(self, *parametric_surfaces, resolution=None, **kwargs):
         self.resolution = (0, 0) if resolution is None else resolution
         super().__init__(uv_func=None, **kwargs)
@@ -308,7 +309,7 @@ class OpenGLSurfaceGroup(OpenGLSurface):
         pass  # Needed?
 
 
-class OpenGLTexturedSurface(OpenGLSurface):
+class TexturedSurface(Surface):
     shader_dtype = [
         ("point", np.float32, (3,)),
         ("du_point", np.float32, (3,)),
@@ -326,8 +327,8 @@ class OpenGLTexturedSurface(OpenGLSurface):
     ):
         self.uniforms = {}
 
-        if not isinstance(uv_surface, OpenGLSurface):
-            raise Exception("uv_surface must be of type OpenGLSurface")
+        if not isinstance(uv_surface, Surface):
+            raise Exception("uv_surface must be of type Surface")
         # Set texture information
         if dark_image_file is None:
             dark_image_file = image_file
@@ -385,3 +386,49 @@ class OpenGLTexturedSurface(OpenGLSurface):
         self.read_data_to_shader(shader_data, "opacity", "opacity")
         self.read_data_to_shader(shader_data, "im_coords", "im_coords")
         return shader_data
+
+class SurfaceMesh(VGroup):
+    def __init__(
+        self,
+        uv_surface,
+        resolution=None,
+        stroke_width=1,
+        normal_nudge=1e-2,
+        depth_test=True,
+        flat_stroke=False,
+        **kwargs
+    ):
+        if not isinstance(uv_surface, Surface):
+            raise Exception("uv_surface must be of type Surface")
+        self.uv_surface = uv_surface
+        self.resolution = resolution if resolution is not None else (21, 21)
+        self.normal_nudge = normal_nudge
+        super().__init__(
+            stroke_width=stroke_width,
+            depth_test=depth_test,
+            flat_stroke=flat_stroke,
+            **kwargs
+        )
+
+    def init_points(self):
+        uv_surface = self.uv_surface
+
+        full_nu, full_nv = uv_surface.resolution
+        part_nu, part_nv = self.resolution
+        u_indices = np.linspace(0, full_nu, part_nu).astype(int)
+        v_indices = np.linspace(0, full_nv, part_nv).astype(int)
+
+        points, du_points, dv_points = uv_surface.get_surface_points_and_nudged_points()
+        normals = uv_surface.get_unit_normals()
+        nudged_points = points + self.normal_nudge * normals
+
+        for ui in u_indices:
+            path = VMobject()
+            full_ui = full_nv * ui
+            path.set_points_smoothly(nudged_points[full_ui : full_ui + full_nv])
+            self.add(path)
+        for vi in v_indices:
+            path = VMobject()
+            path.set_points_smoothly(nudged_points[vi::full_nv])
+            self.add(path)
+
